Add more linting rules && start restructuring the gpt proxy project
