+++ gpt-proxy/src/index.ts
@@ -1,5 +1,5 @@
-import express, { Request, Response, NextFunction } from 'express';
-import { chatWithGpt } from './gpt-prompt.service';
+import express, {Request, Response, NextFunction} from 'express';
+import {chatWithGpt} from './services/gpt-prompt.service';
 import dotenv from 'dotenv';
 
 dotenv.config();
@@ -9,7 +9,7 @@ const app = express();
 // Parse JSON bodies (as sent by API clients)
 app.use(express.json());
 
-app.post('/chat', async (req: Request, res: Response, next: NextFunction) => {
+app.post('/', async (req: Request, res: Response, next: NextFunction) => {
   try {
     const userInput = req.body.content;
 
@@ -22,7 +22,7 @@ app.post('/chat', async (req: Request, res: Response, next: NextFunction) => {
         couchDBResponse: response?.couchSaveResponse
       });
     } else {
-      res.status(400).json({ error: 'Bad Request', message: 'The "content" field must be a non-empty string.' });
+      res.status(400).json({error: 'Bad Request', message: 'The "content" field must be a non-empty string.'});
     }
   } catch (error) {
     next(error);
@@ -30,7 +30,7 @@ app.post('/chat', async (req: Request, res: Response, next: NextFunction) => {
 });
 
 app.use((err: Error, req: Request, res: Response) => {
-  res.status(500).json({ error: 'Internal Server Error', message: err.message });
+  res.status(500).json({error: 'Internal Server Error', message: err.message});
 });
 
 const port = process.env.PORT || 5000;
