+++ gpt-proxy/src/utils/nano-couchdb.service.ts
@@ -1,5 +1,5 @@
-import nano, { DocumentInsertResponse } from 'nano';
-import { ChatItem, iChat } from '../shared/chat.model';
+import nano, {DocumentInsertResponse} from 'nano';
+import {ChatItem, iChat} from '../models/chat.model';
 
 const nanoDB = nano('http://localhost:2200/');
 const db = nanoDB.use('chat_history');
